Change license apache2 (#1233)
* Change license in COPYING, c/c++ code

Replace the commented license notice in c/c++ userspace code with the
apache 2 license, and change the top level COPYING license to apache 2.

Still need to change other code e.g. lua, shell scripts, as well as
changing the license for the driver.

* Change lua code from gpl to apache license

Change lua code that had a clear gpl comment block at the beginning of
the file to apache 2. This doesn't yet cover all files, but the
exceptions will be handled next.

* Update scripts to use apache 2.0 license.

* Update chisels COPYING to Apache 2 license.

* Change copyright for Luca-written files

He was a sysdig employee at the time.

* Add apache copyright to files that didn't have one

For files that had no copyright notice, add the apache copyright notice.

* Use correct copyright years.

* Change Brendan Gregg chisels to apache 2 license

We obtained his permission to change the license to apache 2.0.

* Add dba sysdig to copyright notices

* Change driver license to dual MIT + GPL

In the driver directory, include both the MIT and GPLv2 licenses. In
each file, add a notice that the code is dual licensed under both MIT
and GPLv2.

Update the README to note that the driver is dual-licensed while
the userspace programs are apache 2 licensed.

* Update remaining chisels to use apache license

We got permission from the author via email.

* Change new .c files to apache license

Were recently added and picked up during rebase.