Rewrite the probe builder
The old probe builder used whatever gcc happened to be installed
on the build host, which led to incompatibilities with very old
or very new kernels.

The new builder moves the actual build process to a Docker container
that is chosen based on the actual kernel to build. It also features
a more flexible command line interface (run it without parameters
to see a help message).

A compatibility wrapper for tools using the old script is provided.