Rewrite the probe builder in Python
This requires either a `pip install` from the probe builder directory
or (more likely) a Docker container.

The CLI is slightly changed: to actually build the probes, you need
to add the `build` keyword before the command line arguments.

To run the kernel crawler, run the builder with `crawl <distro>`.

The Docker entrypoint is adapted and usage in a container should not
require any changes.