fix handling of missing kernel packages
some debian kernel packages might disappear from the server
while still being listed in the repository metadata.
This causes an exception in the downloader thread, which will
abort the thread, leaving the main thread stuck while
waiting for the queue to join.
Add some exception handler to the downlaoder thread so that
the caller doesn't get stuck.