batch crawled packages as if it were local
debian and ubuntu have a very weird package naming which
requires us to discard the versioning information our crawler
obtains from the repository metadata, and recur to parsing
filenames, which is what we would be doing on a local distro
anyway.
Notice how this also fixes unpack_kernels for debian, which
expects a list of tuples instead of the dictionary returned
by the crawler.