debian: fix parsing of kernel package version
some debian kernel packages have an additional version suffix
(examples I found are -rt and -cloud)
which is somehow buried into the name, and that is particularly
weird for the the ones having the "common" arch in the name:

linux-headers-5.16.0-1-amd64_5.16.7-2_amd64.deb
linux-headers-5.16.0-1-cloud-amd64_5.16.7-2_amd64.deb
linux-headers-5.16.0-1-common_5.16.7-2_all.deb
linux-headers-5.16.0-1-common-rt_5.16.7-2_all.deb
linux-headers-5.16.0-1-rt-amd64_5.16.7-2_amd64.deb

Refactor the regexp so to capture the combination of arch and variant
(in whatever order) by stopping at the first "_", and handle the
exception of the "common" pseudo-architecture accordingly.

Notice how we now have different combinations of variant+architecture
(or just architecture with no variant) but they are all treated as
if they were an architecture (with the exception of possibly having
an hypen in between, like cloud-amd64).
For this reason, use a different separator ":" between the version
and the variant+architecture so that the version can be
easily reconstructed.