fix building of ubuntu's dishomogenous versions
There's one specific ubuntu kernel, namely 5.15.0-1001-gke,
which comes off as a set of packages which all have a
common version 5.15.0-1001.1, except for
linux-image-5.15.0-1001-gke which has version 5.15.0-1001.1+2.
This is somewhat an exception as normally, when you resolve
packages within the same repo, all kernel packages have
the exact same version, and this is the logic applied by
batch_packages() and unpack_kernels().

However, since we have repo metadata available, we can also leverage
that information, e.g.

  === 5.15.0-1001/1 ===
  http://security.ubuntu.com/ubuntu/pool/main/l/linux-gke/linux-gke-headers-5.15.0-1001_5.15.0-1001.1_amd64.deb
  http://security.ubuntu.com/ubuntu/pool/main/l/linux-signed-gke/linux-image-5.15.0-1001-gke_5.15.0-1001.1+2_amd64.deb
  http://security.ubuntu.com/ubuntu/pool/main/l/linux-gke/linux-modules-5.15.0-1001-gke_5.15.0-1001.1_amd64.deb
  http://security.ubuntu.com/ubuntu/pool/main/l/linux-gke/linux-headers-5.15.0-1001-gke_5.15.0-1001.1_amd64.deb

so that it only needs to be checked that the expected version is at least a prefix
of the package version.
Same goes for unpack_kernels() -- which only works because the "longest" package
does not happen to be the first one, and we surely hope this will not change.

Note: this can be tested by running:

...build ... -k Ubuntu -f 5.15.0-1001